+++ packages/redux-routine/src/test/index.js
@@ -50,6 +50,7 @@ describe( 'createMiddleware', () => {
 	} );
 
 	it( 'should throw if promise rejects', async () => {
+		expect.hasAssertions();
 		const middleware = createMiddleware( {
 			WAIT_FAIL: () => new Promise( ( resolve, reject ) =>
 				reject( 'Message' )
@@ -60,7 +61,7 @@ describe( 'createMiddleware', () => {
 			try {
 				yield { type: 'WAIT_FAIL' };
 			} catch ( error ) {
-				expect( error.message ).toBe( 'Message' );
+				expect( error ).toBe( 'Message' );
 			}
 		}
 
@@ -68,6 +69,7 @@ describe( 'createMiddleware', () => {
 	} );
 
 	it( 'should throw if promise throws', () => {
+		expect.hasAssertions();
 		const middleware = createMiddleware( {
 			WAIT_FAIL: () => new Promise( () => {
 				throw new Error( 'Message' );
@@ -83,6 +85,31 @@ describe( 'createMiddleware', () => {
 		}
 
 		return store.dispatch( createAction() );
+	} );
+
+	// Currently this test will not error even under conditions producing it but
+	// instead will have an uncaught error/warning printed in the cli console:
+	// - (node:37109) UnhandledPromiseRejectionWarning: TypeError: Cannot read
+	// property 'type' of null (and others)
+	// See this github thread for context:
+	// https://github.com/facebook/jest/issues/3251
+	it( 'should handle a null returned from a caught promise error', () => {
+		expect.hasAssertions();
+		const middleware = createMiddleware( {
+			WAIT_FAIL: () => new Promise( () => {
+				throw new Error( 'Message' );
+			} ),
+		} );
+		const store = createStoreWithMiddleware( middleware );
+		function* createAction() {
+			try {
+				yield { type: 'WAIT_FAIL' };
+			} catch ( error ) {
+				expect( error.message ).toBe( 'Message' );
+				return null;
+			}
+		}
+		store.dispatch( createAction() );
 	} );
 
 	it( 'assigns sync controlled return value into yield assignment', () => {